% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annotLODR}
\alias{annotLODR}
\title{Annotate signal-abundance and ratio-abundance plots with LODR}
\usage{
annotLODR(exDat)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}
}
\description{
Annotate signal-abundance and ratio-abundance plots with LODR
}

