\name{COH.RatTox.ILM.MET.CTL.countTable}
\alias{COH.RatTox.ILM.MET.CTL.countTable}
\docType{data}
\title{
Rat toxicogenomics count data
}
\description{
RNA-Seq count data from Methimazole and Control rat biological replicates
}
\usage{data(COH.RatTox.ILM.MET.CTL.countTable)}
\format{
  A data frame with 16590 observations on the following 7 variables.
  \describe{
    \item{\code{Feature}}{a factor vector of all Endogenous and ERCC transcripts in the experiment}
    \item{\code{MET_1}}{a numeric vector of counts from Methimazole treatment biological replicate 1}
    \item{\code{MET_2}}{a numeric vector of counts from Methimazole treatment biological replicate 2}
    \item{\code{MET_3}}{a numeric vector of counts from Methimazole treatment biological replicate 3}
    \item{\code{CTL_1}}{a numeric vector of counts from Control biological replicate 1}
    \item{\code{CTL_2}}{a numeric vector of counts from Control biological replicate 2}
    \item{\code{CTL_3}}{a numeric vector of counts from Control biological replicate 3}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
