\name{dynRangePlot}
\alias{dynRangePlot}
\title{Produce signal-abundance plot to evaluate dynamic range}
\usage{
dynRangePlot(expDat, errorBars = NULL)
}
\arguments{
  \item{expDat}{list, contains input data and stores
  analysis results}

  \item{errorBars}{logical, if TRUE error bars will be
  provided for replicates}
}
\description{
Produce signal-abundance plot to evaluate dynamic range
}

