\name{Lab10.array.UHRR.HBRR}
\alias{Lab10.array.UHRR.HBRR}
\docType{data}
\title{
Lab 10 UHRR and HBRR Illumina BeadArray data
}
\description{
Unnormalized microarray data from Lab 10 of reference sample interlaboratory study
}
\usage{data(Lab10.array.UHRR.HBRR)}
\format{
  A data frame with 17627 observations on the following 7 variables.
  \describe{
    \item{\code{Feature}}{a factor vector of all Endogenous and ERCC transcripts in the experiment}
    \item{\code{UHRR_3}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 1}
    \item{\code{UHRR_2}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 2}
    \item{\code{UHRR_1}}{a numeric vector of fluorescence intensities from UHRR microarray technical replicate 3}
    \item{\code{HBRR_3}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 1}
    \item{\code{HBRR_2}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 2}
    \item{\code{HBRR_1}}{a numeric vector of fluorescence intensities from HBRR microarray technical replicate 3}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
