\name{Lab5.ILM.UHRR.HBRR.countTable}
\alias{Lab5.ILM.UHRR.HBRR.countTable}
\docType{data}
\title{
Lab 5 UHRR and HBRR RNA-Seq Illumina count data
}
\description{
RNA-Seq count data from UHRR and HBRR interlaboratory study library replicates
}
\usage{data(Lab5.ILM.UHRR.HBRR.countTable)}
\format{
  A data frame with 43919 observations on the following 9 variables.
  \describe{
    \item{\code{Feature}}{a character vector of all Endogenous and ERCC transcripts in the experiment}
    \item{\code{UHRR_1}}{a numeric vector of counts from UHRR library preparation replicate 1}
    \item{\code{UHRR_2}}{a numeric vector of counts from UHRR library preparation replicate 2}
    \item{\code{UHRR_3}}{a numeric vector of counts from UHRR library preparation replicate 3}
    \item{\code{UHRR_4}}{a numeric vector of counts from UHRR library preparation replicate 4}
    \item{\code{HBRR_1}}{a numeric vector of counts from HBRR library preparation replicate 1}
    \item{\code{HBRR_2}}{a numeric vector of counts from HBRR library preparation replicate 2}
    \item{\code{HBRR_3}}{a numeric vector of counts from HBRR library preparation replicate 3}
    \item{\code{HBRR_4}}{a numeric vector of counts from HBRR library preparation replicate 4}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
