\name{Lab5.ILM.UHRR.HBRR.totalReads}
\alias{Lab5.ILM.UHRR.HBRR.totalReads}
\docType{data}
\title{
Lab 5 UHRR and HBRR sample total read data
}
\description{
Total reads per library replicate from FASTQ files
}
\usage{data(Lab5.ILM.UHRR.HBRR.totalReads)}
\format{
  The format is:
 int [1:8] 138786892 256006510 199468322 431933806 247985592 219383270 251265814 257508210
}
\details{

}
\source{

}
\references{

}
\examples{
}
\keyword{datasets}
