% Generated by roxygen2 (4.0.0): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiplot function from R cookbook}
\usage{
multiplot(..., plotlist = NULL, cols)
}
\arguments{
\item{...}{comma separated list of plots to include}

\item{plotlist}{list, naming plots to include}

\item{cols}{number of columns for grid of plots}
}
\description{
Multiplot function from R cookbook
}

