% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annotLODR}
\alias{annotLODR}
\title{Annotate signal-abundance and ratio-abundance plots with LODR}
\usage{
annotLODR(exDat)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}
}
\description{
Annotate signal-abundance and ratio-abundance plots with LODR
}
\examples{
data(SEQC.Example)

exDat <- initDat(datType="count", isNorm=FALSE, exTable=MET.CTL.countDat,
                 filenameRoot="testRun", sample1Name="MET",
                 sample2Name="CTL", erccmix="RatioPair",
                 erccdilution=1/100, spikeVol=1, totalRNAmass=0.500,
                 choseFDR=0.1)

exDat <- est_r_m(exDat)

exDat <- dynRangePlot(exDat)
\dontrun{
## this part of example not run during installation
exDat <- geneExprTest(exDat)

exDat <- estLODR(exDat, kind="ERCC", prob=0.9)

exDat <- annotLODR(exDat)

exDat$Figures$maPlot
}
}

