% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dynRangePlot}
\alias{dynRangePlot}
\title{Produce signal-abundance plot to evaluate dynamic range}
\usage{
dynRangePlot(exDat)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}
}
\description{
Produce signal-abundance plot to evaluate dynamic range
}

