\name{erccROC}
\alias{erccROC}
\title{Produce Receiver Operator Characteristic (ROC) Curves and AUC statistics}
\usage{
erccROC(expDat, pValDat = NULL)
}
\arguments{
  \item{expDat}{list, contains input data and stores
  analysis results}

  \item{pValDat}{string, Pval csv file name, default is
  NULL}
}
\description{
Produce Receiver Operator Characteristic (ROC) Curves and
AUC statistics
}

