% Generated by roxygen2 (4.0.1): do not edit by hand
\name{erccROC}
\alias{erccROC}
\title{Produce Receiver Operator Characteristic (ROC) Curves and AUC statistics}
\usage{
erccROC(exDat)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}
}
\description{
Produce Receiver Operator Characteristic (ROC) Curves and AUC statistics
}

