% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maSignal}
\alias{maSignal}
\title{Generate MA plots with or without annotation using LODR estimates}
\usage{
maSignal(exDat, alphaPoint = 0.8, r_mAdjust = TRUE, replicate = TRUE)
}
\arguments{
\item{exDat}{list, contains input data and stores analysis results}

\item{alphaPoint}{numeric value, for alpha (transparency) for plotted points,
range is 0 - 1}

\item{r_mAdjust}{default is TRUE, if FALSE then the r_m estimate will not
used to offset dashed lines for empirical ratios on figure}

\item{replicate}{default is TRUE, if FALSE then error bars will not be
                   produced}
}
\description{
Generate MA plots with or without annotation using LODR estimates
}
\examples{
\donttest{
data(SEQC.Example)

exDat <- initDat(datType="count", isNorm = FALSE, exTable=MET.CTL.countDat,
                 filenameRoot = "testRun",sample1Name = "MET",
                 sample2Name = "CTL", erccmix = "RatioPair",
                 erccdilution = 1/100, spikeVol = 1, totalRNAmass = 0.500,
                 choseFDR = 0.1)

exDat <- est_r_m(exDat)

exDat <- dynRangePlot(exDat)

exDat <- geneExprTest(exDat)
# generate MA plot without LODR annotation
exDat <- maSignal(exDat)

exDat$Figures$maPlot

# Include LODR annotation
exDat <- annotLODR(exDat)

exDat$Figures$maPlot
}
}

